+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # CHANGELOG
 
 ## [Unreleased]
-
+### Added
+* Added `Kreait\Firebase\RemoteConfig\DefaultValue` now has an added `value()` method to 
+  retrieve a default value's value.
+  
 ## [5.6.0] - 2020-07-02
 ### Added
 * User Records now contain the date and time when a user's password has last been updated.